+++ packages/api/src/runner-lcd.ts
@@ -1,4 +1,5 @@
 import LCDClient from './util/lcd'
+import { IMsg } from './transaction'
 
 export type IRunner = {
   hash: string;
@@ -6,7 +7,40 @@ export type IRunner = {
   instanceHash?: string | null;
 }
 
+export type IMsgCreate = {
+  address: string;
+  serviceHash: string;
+  envHash: string;
+}
+
+export type IMsgDelete = {
+  address: string;
+  runnerHash: string;
+}
+
 export default class Runner extends LCDClient {
+
+  createMsg(owner: string, serviceHash: string, envHash: string): IMsg<IMsgCreate> {
+    return {
+      type: 'runner/CreateRunner',
+      value: {
+        address: owner,
+        serviceHash: serviceHash,
+        envHash: envHash
+      }
+    }
+  }
+
+  deleteMsg(owner: string, runnerHash: string): IMsg<IMsgDelete> {
+    return {
+      type: 'runner/DeleteRunner',
+      value: {
+        address: owner,
+        runnerHash: runnerHash
+      }
+    }
+  }
+
   async get(hash: string): Promise<IRunner> {
     return (await this.query(`/runner/get/${hash}`)).result
   }
@@ -16,5 +50,17 @@ export default class Runner extends LCDClient {
     if (filter && filter.instanceHash) runners = runners.filter(x => x.instanceHash === filter.instanceHash)
     if (filter && filter.address) runners = runners.filter(x => x.address === filter.address)
     return runners
+  }
+
+  async hash(owner: string, serviceHash: string, env: string[]): Promise<{
+    runnerHash: string,
+    instanceHash: string,
+    envHash: string
+  }> {
+    return (await this.query('/runner/hash', {
+      serviceHash,
+      env,
+      address: owner,
+    }, 'POST')).result
   }
 }