+++ packages/cli/src/commands/service/dev.ts
@@ -5,7 +5,7 @@ import LCD from '@mesg/api/lib/lcd'
 import API from '@mesg/api'
 import chalk from 'chalk'
 import { decode } from '@mesg/api/lib/util/encoder'
-import { parseLog, listServices } from '../../utils/docker'
+import { parseLog, listContainers } from '../../utils/docker'
 import * as Environment from '../../utils/environment-tasks'
 import * as Service from '../../utils/service'
 import * as Runner from '../../utils/runner'
@@ -73,8 +73,8 @@ export default class Dev extends Command {
       },
       {
         title: 'Starting service',
-        task: async () => {
-          runner = await Runner.create(this.grpc, this.lcd, service.hash, flags.env)
+        task: async ctx => {
+          runner = await Runner.create(this.lcdEndpoint, ctx.mnemonic, service.hash, flags.env)
         }
       },
       {
@@ -83,8 +83,8 @@ export default class Dev extends Command {
           {
             title: 'Fetching service\'s logs',
             task: async () => {
-              const dockerServices = await listServices({ label: [`mesg.runner=${runner.hash}`] })
-              this.logs = await Promise.all(dockerServices.map(x => x.logs({
+              const dockerContainers = await listContainers({ label: [`mesg.runner=${runner.hash}`] })
+              this.logs = await Promise.all(dockerContainers.map(x => x.logs({
                 stderr: true,
                 stdout: true,
                 follow: true,
@@ -119,7 +119,7 @@ export default class Dev extends Command {
         ])
       }
     ])
-    await tasks.run({
+    const res = await tasks.run({
       configDir: this.config.dataDir,
       image: flags.image,
       pull: flags.pull,
@@ -152,6 +152,13 @@ export default class Dev extends Command {
             if (this.events) this.events.cancel()
             if (this.results) this.results.cancel()
             return Promise.resolve()
+          }
+        },
+        {
+          title: 'Stopping service',
+          skip: () => !service && !runner,
+          task: async () => {
+            return Runner.stop(this.lcdEndpoint, res.mnemonic, runner.hash)
           }
         },
         Environment.stop,