+++ packages/cli/CHANGELOG.md
@@ -40,6 +40,8 @@ Commands `service:dev` and `process:dev` are automatically start a local environ
 - [#194](https://github.com/mesg-foundation/js-sdk/pull/194) Add `deploy:service` command
 - [#196](https://github.com/mesg-foundation/js-sdk/pull/196) Update code organization for tasks
 - [#197](https://github.com/mesg-foundation/js-sdk/pull/197) State of the local chain in the `dataDir` (https://oclif.io/docs/config)
+- [#199](https://github.com/mesg-foundation/js-sdk/pull/199) Use api helper to get the hash of a created service/process
+- [#199](https://github.com/mesg-foundation/js-sdk/pull/199) Run engine and service with the `@mesg/runner` library
 
 #### Bug fixes